# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files/upload-intent": {
      "post": {
        "operationId": "FilesController_createUploadIntent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadIntentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload intent with signed URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadIntentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a signed upload URL for a file",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/{id}/mark-uploaded": {
      "patch": {
        "operationId": "FilesController_markUploaded",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "File ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkUploadedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File marked as uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileMarkUploadedResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm a file has been uploaded to storage",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/{id}": {
      "get": {
        "operationId": "FilesController_getFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "File ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File record with signed URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get file metadata and signed download URL",
        "tags": [
          "Files"
        ]
      }
    },
    "/chat/conversations": {
      "post": {
        "operationId": "ChatController_createConversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversation created or existing one returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get or create a conversation with another user",
        "tags": [
          "Chat"
        ]
      },
      "get": {
        "operationId": "ChatController_listConversations",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "Filter conversations by context",
            "schema": {
              "example": "all",
              "type": "string",
              "enum": [
                "all",
                "buy",
                "sell"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of conversations with last message preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationsListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List conversations for the current user",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/conversations/{id}": {
      "get": {
        "operationId": "ChatController_getConversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Conversation ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Not a participant of this conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single conversation with metadata for the current user",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/conversations/{id}/messages": {
      "get": {
        "operationId": "ChatController_listMessages",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of messages to return (1–100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "description": "Cursor — return messages sent before this ISO 8601 timestamp",
            "schema": {
              "example": "2024-06-01T12:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Conversation ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of messages in descending sent_at order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesListResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Not a participant of this conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List messages in a conversation (cursor-paginated)",
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/v2/followups/{itemType}/{itemId}/chat/conversations": {
      "post": {
        "operationId": "KayanController_createConversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFollowupConversationBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Kayan V2"
        ]
      }
    },
    "/v2/followups/{itemType}/{itemId}/chat/conversations/{id}/messages": {
      "get": {
        "operationId": "KayanController_listMessages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Kayan V2"
        ]
      },
      "post": {
        "operationId": "KayanController_sendMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendFollowupMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Kayan V2"
        ]
      }
    },
...
  },
  "info": {
    "title": "Market Place API",
    "description": "Online marketplace REST API — user registration, product listings, real-time chat, ratings, and admin moderation.\n\n> **WebSocket (Chat):** Real-time messaging uses a Socket.io gateway at the `/chat` namespace. That interface is not documented here; see the project README for event names and payloads.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "Conflict"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register/request-otp"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "data": {
            "type": "object",
            "additionalProperties": false,
            "example": null,
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "error"
        ]
      },
...
      "CreateUploadIntentDto": {
        "type": "object",
        "properties": {
          "ownerType": {
            "type": "string",
            "enum": [
              "user",
              "product",
              "message"
            ],
            "description": "Entity type that owns this file",
            "example": "product"
          },
          "ownerId": {
            "type": "number",
            "description": "ID of the owning entity (omit for new entities)",
            "example": 5,
            "minimum": 1
          },
          "purpose": {
            "type": "string",
            "enum": [
              "avatar",
              "product_image",
              "chat_attachment",
              "document"
            ],
            "description": "Intended use of the file",
            "example": "product_image"
          },
          "filename": {
            "type": "string",
            "description": "Original filename including extension",
            "example": "photo.jpg",
            "minLength": 1,
            "maxLength": 255
          },
          "mimeType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp",
              "image/gif",
              "video/mp4",
              "video/quicktime",
              "video/webm",
              "video/x-msvideo",
              "audio/mpeg",
              "audio/wav",
              "audio/ogg",
              "audio/webm",
              "application/pdf",
              "application/msword",
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
              "application/vnd.ms-excel",
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
              "application/zip"
            ],
            "description": "MIME type of the file",
            "example": "image/jpeg"
          },
          "fileSizeBytes": {
            "type": "number",
            "description": "File size in bytes (max 50 MB)",
            "example": 204800,
            "minimum": 0,
            "maximum": 52428800
          }
        },
        "required": [
          "ownerType",
          "purpose",
          "filename",
          "mimeType"
        ]
      },
      "UploadIntentFileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 42
          },
          "objectKey": {
            "type": "string",
            "example": "users/5/avatar.jpg"
          },
          "status": {
            "type": "string",
            "example": "pending",
            "enum": [
              "pending",
              "uploaded",
              "failed",
              "deleted"
            ]
          }
        },
        "required": [
          "id",
          "objectKey",
          "status"
        ]
      },
      "UploadIntentDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "example": "PUT",
            "description": "HTTP method to use for the upload request"
          },
          "url": {
            "type": "string",
            "example": "https://api.cloudinary.com/v1_1/example/upload"
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-03-28T12:15:00.000Z"
          },
          "headers": {
            "type": "object",
            "example": {
              "Content-Type": "image/jpeg"
            },
            "description": "Headers to include in the upload request (optional)"
          },
          "fields": {
            "type": "object",
            "example": {
              "signature": "abc123",
              "timestamp": "1711622400"
            },
            "description": "Form fields for multipart uploads (optional)"
          }
        },
        "required": [
          "method",
          "url",
          "expiresAt"
        ]
      },
      "UploadIntentDataDto": {
        "type": "object",
        "properties": {
          "file": {
            "$ref": "#/components/schemas/UploadIntentFileDto"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadIntentDto"
          }
        },
        "required": [
          "file",
          "upload"
        ]
      },
      "UploadIntentResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/UploadIntentDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "MarkUploadedDto": {
        "type": "object",
        "properties": {
          "checksumSha256": {
            "type": "string",
            "description": "SHA-256 checksum of the uploaded file (hex, 64 chars) for integrity verification",
            "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
            "minLength": 64,
            "maxLength": 64
          }
        }
      },
      "FileMarkUploadedDataDto": {
        "type": "object",
        "properties": {
          "file": {
            "$ref": "#/components/schemas/UploadIntentFileDto"
          }
        },
        "required": [
          "file"
        ]
      },
      "FileMarkUploadedResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/FileMarkUploadedDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "FileMetadataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 42
          },
          "uploader_user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "owner_type": {
            "type": "string",
            "example": "user",
            "enum": [
              "user",
              "product",
              "message"
            ]
          },
          "owner_id": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "purpose": {
            "type": "string",
            "example": "avatar",
            "enum": [
              "avatar",
              "product_image",
              "chat_attachment",
              "document"
            ]
          },
          "object_key": {
            "type": "string",
            "example": "users/5/avatar.jpg"
          },
          "mime_type": {
            "type": "object",
            "example": "image/jpeg",
            "nullable": true
          },
          "file_size_bytes": {
            "type": "object",
            "example": 204800,
            "nullable": true
          },
          "status": {
            "type": "string",
            "example": "uploaded",
            "enum": [
              "pending",
              "uploaded",
              "failed",
              "deleted"
            ]
          },
          "created_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "uploaded_at": {
            "type": "object",
            "example": "2026-03-28T12:05:00.000Z",
            "nullable": true
          },
          "readUrl": {
            "type": "string",
            "example": "https://res.cloudinary.com/example/image/upload/users/5/avatar.jpg"
          }
        },
        "required": [
          "id",
          "owner_type",
          "purpose",
          "object_key",
          "status",
          "created_at",
          "readUrl"
        ]
      },
      "FileDataDto": {
        "type": "object",
        "properties": {
          "file": {
            "$ref": "#/components/schemas/FileMetadataDto"
          }
        },
        "required": [
          "file"
        ]
      },
      "FileResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/FileDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "CreateConversationDto": {
        "type": "object",
        "properties": {
          "participantId": {
            "type": "number",
            "description": "User ID of the other conversation participant",
            "example": 12,
            "minimum": 1
          },
          "productId": {
            "type": "number",
            "description": "Optional product ID to attach as context for the conversation",
            "example": 45
          }
        },
        "required": [
          "participantId"
        ]
      },
...
      "ConversationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "user_a": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "user_b": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "product": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedProductDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "user_a",
          "user_b",
          "created_at"
        ]
      },
      "ConversationDataDto": {
        "type": "object",
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationDto"
          }
        },
        "required": [
          "conversation"
        ]
      },
      "ConversationResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ConversationDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "ConversationWithLastMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "user_a": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "user_b": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "product": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedProductDto"
              }
            ]
          },
          "last_message": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedMessageDto"
              }
            ]
          },
          "last_message_text": {
            "type": "object",
            "example": "Hello, is this still available?",
            "nullable": true
          },
          "last_message_sent_at": {
            "type": "object",
            "example": "2026-03-28T13:00:00.000Z",
            "nullable": true
          },
          "peer_user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "unread_count": {
            "type": "number",
            "example": 2
          },
          "product_name": {
            "type": "object",
            "example": "iPhone 13",
            "nullable": true
          },
          "product_price": {
            "type": "object",
            "example": 600,
            "nullable": true
          },
          "product_image": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedFileDto"
              }
            ]
          },
          "product_image_object_key": {
            "type": "object",
            "example": "products/1/image.jpg",
            "nullable": true
          }
        },
        "required": [
          "id",
          "user_a",
          "user_b",
          "created_at",
          "peer_user",
          "unread_count"
        ]
      },
      "ConversationsListDataDto": {
        "type": "object",
        "properties": {
          "conversations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationWithLastMessageDto"
            }
          }
        },
        "required": [
          "conversations"
        ]
      },
      "ConversationsListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ConversationsListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "MessagesListDataDto": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDto"
            }
          }
        },
        "required": [
          "messages"
        ]
      },
      "MessagesListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/MessagesListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "CreateFollowupConversationBodyDto": {
        "type": "object",
        "properties": {}
      },
      "SendFollowupMessageDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}