# src/chat/chat.controller.ts
import { Body, Controller, Get, Param, Post, Query, UseGuards } from '@nestjs/common';
import { ApiBearerAuth, ApiOperation, ApiParam, ApiResponse, ApiTags } from '@nestjs/swagger';
import { CurrentUser } from '../common/decorators/current-user.decorator';
import { JwtAuthGuard } from '../common/guards/jwt-auth.guard';
import { AuthUser } from '../common/types/auth-user.type';
import { ErrorResponseDto } from '../common/dto/error-response.dto';
import { ChatService } from './chat.service';
import { ChatJoinedPayloadBuilder } from './chat-joined-payload.builder';
import { ChatSocketRegistryService } from './chat-socket-registry.service';
import { CreateConversationDto } from './dto/create-conversation.dto';
import { ListConversationsDto } from './dto/list-conversations.dto';
import { ListMessagesDto } from './dto/list-messages.dto';
import {
  ConversationResponseDto,
  ConversationsListResponseDto,
  MessagesListResponseDto,
} from './dto/chat-response.dto';
import { IdParamDto } from '../common/dto/id-param.dto';

@ApiTags('Chat')
@ApiBearerAuth()
@Controller('chat')
@UseGuards(JwtAuthGuard)
export class ChatController {
  constructor(
    private readonly chatService: ChatService,
    private readonly chatSocketRegistry: ChatSocketRegistryService,
    private readonly chatJoinedPayloadBuilder: ChatJoinedPayloadBuilder,
  ) {}

  @Post('conversations')
  @ApiOperation({ summary: 'Get or create a conversation with another user' })
  @ApiResponse({ status: 201, description: 'Conversation created or existing one returned', type: ConversationResponseDto })
  async createConversation(
    @CurrentUser() user: AuthUser,
    @Body() dto: CreateConversationDto,
  ): Promise<Record<string, unknown>> {
    const response = await this.chatService.getOrCreateConversation(user.sub, dto.participantId, dto.productId);
    const conversation = (response as { conversation?: Record<string, unknown> }).conversation;
    const conversationId = Number(conversation?.id);
    if (Number.isInteger(conversationId) && conversationId > 0) {
      const participants = await this.chatService.getConversationParticipants(conversationId);
      const payload = await this.chatJoinedPayloadBuilder.buildConversationJoinedPayload(response);
      await this.chatSocketRegistry.emitConversationJoinedToParticipants(
        conversationId,
        [participants.userAId, participants.userBId],
        payload,
      );
    }
    return response;
  }

  @Get('conversations')
  @ApiOperation({ summary: 'List conversations for the current user' })
  @ApiResponse({ status: 200, description: 'Array of conversations with last message preview', type: ConversationsListResponseDto })
  listConversations(
    @CurrentUser() user: AuthUser,
    @Query() query: ListConversationsDto,
  ): Promise<Record<string, unknown>> {
    return this.chatService.listConversations(user.sub, query.scope ?? 'all', query.limit ?? 20, query.offset ?? 0);
  }

  @Get('conversations/:id')
  @ApiParam({ name: 'id', type: Number, description: 'Conversation ID' })
  @ApiOperation({ summary: 'Get a single conversation with metadata for the current user' })
  @ApiResponse({ status: 200, description: 'Conversation metadata', type: ConversationResponseDto })
  @ApiResponse({ status: 403, description: 'Not a participant of this conversation', type: ErrorResponseDto })
  @ApiResponse({ status: 404, description: 'Conversation not found', type: ErrorResponseDto })
  getConversation(
    @CurrentUser() user: AuthUser,
    @Param() params: IdParamDto,
  ): Promise<Record<string, unknown>> {
    return this.chatService.getConversationById(user.sub, params.id);
  }

  @Get('conversations/:id/messages')
  @ApiParam({ name: 'id', type: Number, description: 'Conversation ID' })
  @ApiOperation({ summary: 'List messages in a conversation (cursor-paginated)' })
  @ApiResponse({ status: 200, description: 'Array of messages in descending sent_at order', type: MessagesListResponseDto })
  @ApiResponse({ status: 403, description: 'Not a participant of this conversation', type: ErrorResponseDto })
  @ApiResponse({ status: 404, description: 'Conversation not found', type: ErrorResponseDto })
  listMessages(
    @CurrentUser() user: AuthUser,
    @Param() params: IdParamDto,
    @Query() dto: ListMessagesDto,
  ): Promise<Record<string, unknown>> {
    return this.chatService.listMessages(user.sub, params.id, dto.limit ?? 20, dto.before);
  }
}